backups/SS_Backup_20251122_1223/backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
python-jose[cryptography]
passlib[bcrypt]
python-multipart
yfinance
